Created Navbar Partial
Created a nav bar partial using the bootstrap navbar sample code.
Partials are created by making a new .html.erb file in views/layouts and starting its name with an underscore.
This navbar has rails links that involve a new to me tag `class`. This will allow it to inherit css styling and is formatted as class:'class-name'